requirements.txt
requests
python-dotenv
beautifulsoup4
lxml
matplotlib
